backend/requirements.txt
google-genai
beautifulsoup4==4.12.3
requests==2.31.0
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
